renamed a per new node name
