+++ CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 3.15.3)
-project(win_camera)
+project(msft_camera)
 
 find_package(catkin REQUIRED COMPONENTS
   image_transport roscpp sensor_msgs nodelet camera_info_manager roslint std_msgs message_generation roslint)
@@ -35,22 +35,22 @@ include_directories(
 ## Declare a cpp library
 add_library(win_camera STATIC src/winrospublisher.cpp src/wincapture.cpp)
 
-add_dependencies(win_camera win_camera_generate_messages_cpp)
+add_dependencies(win_camera msft_camera_generate_messages_cpp)
 
-add_library(win_camera_nodelet SHARED src/wincameranodelet.cpp)
-set_target_properties(win_camera_nodelet PROPERTIES WINDOWS_EXPORT_ALL_SYMBOLS TRUE) 
+add_library(msft_camera_nodelet SHARED src/msftcameranodelet.cpp)
+set_target_properties(msft_camera_nodelet PROPERTIES WINDOWS_EXPORT_ALL_SYMBOLS TRUE) 
 
 ## Declare a cpp executable
-add_executable(win_camera_node src/wincameranode.cpp)
-add_dependencies(win_camera_node win_camera)
-add_dependencies(win_camera_nodelet win_camera)
+add_executable(msft_camera_node src/msftcameranode.cpp)
+add_dependencies(msft_camera_node win_camera)
+add_dependencies(msft_camera_nodelet win_camera)
 
 ## Specify libraries to link a library or executable target against
 target_link_libraries(win_camera
   ${catkin_LIBRARIES}
   )
 
-target_link_libraries(win_camera_nodelet
+target_link_libraries(msft_camera_nodelet
   ${catkin_LIBRARIES}
   win_camera
   mf
@@ -62,7 +62,7 @@ target_link_libraries(win_camera_nodelet
   runtimeobject
   )
 
-target_link_libraries(win_camera_node
+target_link_libraries(msft_camera_node
   ${catkin_LIBRARIES}
   win_camera
   mf
@@ -86,7 +86,7 @@ target_link_libraries(win_camera_node
 # )
 
 ## Mark executables and/or libraries for installation
-install(TARGETS win_camera win_camera_nodelet win_camera_node
+install(TARGETS win_camera msft_camera_nodelet msft_camera_node
   ARCHIVE DESTINATION ${CATKIN_PACKAGE_LIB_DESTINATION}
   LIBRARY DESTINATION ${CATKIN_PACKAGE_LIB_DESTINATION}
   RUNTIME DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION}
@@ -97,7 +97,7 @@ install(DIRECTORY include
   )
 
 # add xml file
-install(FILES win_camera_nodelets.xml
+install(FILES msft_camera_nodelets.xml
   DESTINATION ${CATKIN_PACKAGE_SHARE_DESTINATION}
   )
 
@@ -107,16 +107,16 @@ install(FILES win_camera_nodelets.xml
 #############
 
 roslint_cpp(
-  src/wincameranode.cpp src/wincameranodelet.cpp src/wincapture.cpp src/winrospublisher.cpp
+  src/msftcameranode.cpp src/msftcameranodelet.cpp src/wincapture.cpp src/winrospublisher.cpp
   )
 
 if (CATKIN_ENABLE_TESTING)
   find_package(rostest REQUIRED)
-  add_rostest_gtest(testWinCamera test/winCamera.test
-    test/testWinCamera.cpp)
-  target_link_libraries(testWinCamera ${PROJECT_NAME} ${catkin_LIBRARIES})
+  add_rostest_gtest(testMsftCamera test/msftCamera.test
+    test/testMsftCamera.cpp)
+  target_link_libraries(testMsftCamera win_camera ${catkin_LIBRARIES})
 
-  add_rostest_gtest(testWinCameraNoYaml test/WinCameraNoYaml.test
-    test/testWinCameraNoYaml.cpp)
-  target_link_libraries(testWinCameraNoYaml ${PROJECT_NAME} ${catkin_LIBRARIES})
+  add_rostest_gtest(testMsftCameraNoYaml test/msftCameraNoYaml.test
+    test/testMsftCameraNoYaml.cpp)
+  target_link_libraries(testMsftCameraNoYaml win_camera ${catkin_LIBRARIES})
 endif()
