+++ src/winrospublisher.cpp
@@ -1,6 +1,6 @@
 // Copyright (C) Microsoft Corporation. All rights reserved.
 #include "winrospublisher.h"
-namespace ros_win_camera
+namespace ros_msft_camera
 {
 
 }
