--- test/WinCameraNoYaml.test
@@ -1,9 +1,0 @@
-<launch>
-  <node pkg="win_camera" type="win_camera_node" name="win_camera_no_yaml">
-    <param name="frame_id" value="WinCamera2" />
-    <param name="image_width" value="1280" />
-    <param name="image_height" value="720" />
-    <param name="frame_rate" value="30.0" />
-  </node>
-  <test test-name="testWinCameraNoYaml" pkg="win_camera" type="testWinCameraNoYaml" />
-</launch>
