+++ test/msftCamera.test
@@ -3,6 +3,8 @@
     <param name="frame_id" value="MsftCamera1" />
     <param name="image_width" value="1280" />
     <param name="image_height" value="720" />
+    <param name="camera_info_url" value="file://$(find msft_camera)\test\sample.yaml" />
     <param name="frame_rate" value="30.0" />
   </node>
+  <test test-name="testMsftCamera" pkg="msft_camera" type="testMsftCamera" />
 </launch>
