+++ test/msftCameraNoYaml.test
@@ -1,0 +1,9 @@
+<launch>
+  <node pkg="msft_camera" type="msft_camera_node" name="msft_camera_no_yaml">
+    <param name="frame_id" value="MsftCamera2" />
+    <param name="image_width" value="1280" />
+    <param name="image_height" value="720" />
+    <param name="frame_rate" value="30.0" />
+  </node>
+  <test test-name="testMsftCameraNoYaml" pkg="msft_camera" type="testMsftCameraNoYaml" />
+</launch>
