+++ test/sample.yaml
@@ -1,6 +1,6 @@
-image_width: 2448
-image_height: 2050
-camera_name: WinCamera1
+image_width: 1280
+image_height: 720
+camera_name: MsftCamera1
 camera_matrix:
   rows: 3
   cols: 3
